refactor: rename Rules to ParserRules
This makes it consistent with the database.
